% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_call_cnv.R
\name{batch_call_cnv}
\alias{batch_call_cnv}
\title{Call CNVs from read count files}
\usage{
batch_call_cnv(
  cnm_path,
  cnn_paths,
  cnn_dirs,
  out_dir,
  test_id,
  ref_ids,
  by_gender = FALSE,
  fraction = c(0.5, 0.5)
)
}
\arguments{
\item{cnm_path}{Path of merged CNN file.}

\item{cnn_paths}{Path of CNN files, separated by comma(,).}

\item{cnn_dirs}{Path of CNN directory, separated by comma(,).}

\item{out_dir}{Output directory path.}

\item{test_id}{Sample id to call CNV.}

\item{ref_ids}{Reference ids to create a baseline, separated by comma(,).}

\item{by_gender}{A bool value to call by gender.}

\item{fraction}{Fraction of CNV >=100Kb and >=5Mb, default:c(0.5, 0.5).}
}
\description{
Call CNVs from read count files
}
