% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_cov.R
\name{batch_cov}
\alias{batch_cov}
\title{Calculate read count in bams}
\usage{
batch_cov(bam_dirs, bam_paths, bed_path, out_dir, thread = 4, mapq.filter = 0)
}
\arguments{
\item{bam_dirs}{Path of BAM directory, separated by comma(,).}

\item{bam_paths}{Path of BAM file, separated by comma(,).}

\item{bed_path}{Path of BED file.}

\item{out_dir}{A character string of directory to output coverage files.}

\item{thread}{An integer providing the number of thread, default: 4.}

\item{mapq.filter}{A non-negative integer specifying the minimum mapping
quality to include. BAM reads with mapping qualities less than mapqFilter
are discarded, default: 0.}
}
\description{
Calculate read count in bams
}
