% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_gender.R
\name{batch_gender}
\alias{batch_gender}
\title{Calculate gender and total read from a coverage file.}
\usage{
batch_gender(cnn_paths, cnn_dirs, out_dir)
}
\arguments{
\item{cnn_paths}{Path of CNN files, separated by comma(,).}

\item{cnn_dirs}{Path of CNN directory, separated by comma(,).}

\item{out_dir}{Output directory path.}
}
\description{
Calculate gender and total read from a coverage file.
}
