% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_merge.R
\name{batch_merge}
\alias{batch_merge}
\title{Merge CNN, VCF, CNR, CNS file.}
\usage{
batch_merge(cnn_paths, cnn_dirs, auto_file, sex_file, out_path)
}
\arguments{
\item{cnn_paths}{Path of CNN files, separated by comma(,).}

\item{cnn_dirs}{Path of CNN directory, separated by comma(,).}

\item{auto_file}{path of VCF, CNR, CNS file.}

\item{sex_file}{path of VCF, CNR, CNS file.}

\item{out_path}{Output file path.}
}
\description{
Input cnn_paths or cnn_dirs, it will merge multiple samples'
  read count result files(.cnn) and create a merged file. Input auto_file and
  sex_file, it will merge autosome and sex chromosome result, the VCF, CNR,
  CNS file can be input.
}
