% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_split.R
\name{batch_split}
\alias{batch_split}
\title{Split wgs region into bins}
\usage{
batch_split(path, bin = 10000, access_bed = "")
}
\arguments{
\item{path}{Path of output, a BED file.}

\item{bin}{An integer of bin size, default: 1E4.}

\item{access_bed}{BED file include sequence-accessible region. If NULL it
will use whole genome region.}
}
\description{
Split wgs region into bins
}
