% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_cnv.R
\name{call_cnv}
\alias{call_cnv}
\title{Calling CNV}
\usage{
call_cnv(data, s_id, is_male = TRUE, fraction = c(0.5, 0.5))
}
\arguments{
\item{data}{A data frame of log2ratio to do CBS. It must has three column:
chromosome, end, log2.}

\item{s_id}{A charactor string of sample id.}

\item{is_male}{A bool value, TRUE when the gender is male.
change-points, default: 0.05.}

\item{fraction}{Fraction of CNV >=100Kb and >=5Mb, default:c(0.5, 0.5).}
}
\description{
Calling CNV
}
