% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV.R
\name{CNV}
\alias{CNV}
\title{CNV}
\usage{
CNV(file = NA, Chromosome = NA)
}
\arguments{
\item{file}{Provide text file produced by bin-by-sam_2.0.py}

\item{Specify}{a particular chromosome or a vector of Chromosomes}
}
\description{
Plots Copy Number Variants after Sam files have been processed by bin-by-sam_2.0.py
}
