% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_reads_of_region.R
\name{count_reads_of_region}
\alias{count_reads_of_region}
\title{Calculating read count in region}
\usage{
count_reads_of_region(region, bamPath, mapq.filter = 0, minoverlap = 75L)
}
\arguments{
\item{region}{A grange object of region to extract reads in BAM file.}

\item{bamPath}{A character string of the BAM path.}

\item{mapq.filter}{A non-negative integer specifying the minimum mapping
quality to include. BAM reads with mapping qualities less than mapqFilter
are discarded.}

\item{minoverlap}{Minimum overlap size for region and reads, default: 75L.}
}
\description{
Calculating read count in region
}
