% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creat_chunk.R
\name{creat_chunk}
\alias{creat_chunk}
\title{Create grange into chunk}
\usage{
creat_chunk(gr, bin = 10000)
}
\arguments{
\item{gr}{A grange object.}

\item{bin}{Bin size, default: 1E4.}
}
\description{
Create grange into chunk
}
