% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_cbs.R
\name{do_cbs}
\alias{do_cbs}
\title{Circular Binary Segmentation}
\usage{
do_cbs(data, s_id, alpha = 0.05, min.width = 2)
}
\arguments{
\item{data}{A data frame of log2ratio to do CBS. It must has three column:
chromosome, end, log2.}

\item{s_id}{A charactor string of sample id.}

\item{alpha}{A numeric value of significance levels for the test to accept
change-points, default: 0.05.}

\item{min.width}{An integer value of the minimum number of markers for a
changed segment, default: 2.}
}
\description{
Circular Binary Segmentation
}
