% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_id.R
\name{get_id}
\alias{get_id}
\title{Extract sample id from file path}
\usage{
get_id(file, ptn = NULL)
}
\arguments{
\item{file}{Path of file.}

\item{ptn}{A charactor for spliting string. By default NULL, it will split
string by '.'.}
}
\description{
Extract sample id from file path
}
