% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wgs_bin.R
\name{get_wgs_bin}
\alias{get_wgs_bin}
\title{Split wgs region into bins}
\usage{
get_wgs_bin(bin = 10000, access_bed = "")
}
\arguments{
\item{bin}{An integer of bin size, default: 1E4.}

\item{access_bed}{BED file include sequence-accessible region. If NULL it
will use whole genome region.}
}
\description{
Split wgs region into bins
}
