% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_bin_df.R
\name{merge_bin_df}
\alias{merge_bin_df}
\title{Merge log2ratio with a specific interval of bins}
\usage{
merge_bin_df(data, group_size = 50)
}
\arguments{
\item{data}{A data frame of cnr file.}

\item{group_size}{A interval to merge, default: 50.}
}
\description{
Merge log2ratio with a specific interval of bins
}
