% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_result_files.R
\name{merge_result_files}
\alias{merge_result_files}
\title{Merge autosome and sex chromosome result}
\usage{
merge_result_files(auto_file, sex_file, out_path)
}
\arguments{
\item{auto_file}{path of VCF, CNR, CNS file.}

\item{sex_file}{path of VCF, CNR, CNS file.}

\item{out_path}{path of output file.}
}
\description{
Merge autosome and sex chromosome result
}
