% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merger_cov_files.R
\name{merger_cov_files}
\alias{merger_cov_files}
\title{Merge coverage files.}
\usage{
merger_cov_files(files, path = NULL)
}
\arguments{
\item{files}{A character vector contains several coverage files path.}

\item{path}{Path to write to.}
}
\value{
A data frame, of which columns are chromosome, start position, end
  position, and depths in input coverage files.
}
\description{
Align and merge coverage files (<filename>.cnn) with chromosome,
  start and end position. Four required fields in a coverage file are
  chromosome name, start and end position, depth of coverage.
}
