% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_vcf.R
\name{output_vcf}
\alias{output_vcf}
\title{Output CNV calling result to VCF file}
\usage{
output_vcf(dat, path)
}
\arguments{
\item{dat}{A data.frame of CNV calling result.}

\item{path}{Path of VCF file.}
}
\description{
Output CNV calling result to VCF file
}
