% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidy_to_cn.R
\name{ploidy_to_cn}
\alias{ploidy_to_cn}
\title{Count copy numer with different mosaicism level}
\usage{
ploidy_to_cn(f, ploidy, cn)
}
\arguments{
\item{f}{Fraction of variants.}

\item{ploidy}{the ploidy of chromosome.}

\item{cn}{Copy number.}
}
\description{
Count copy numer with different mosaicism level
}
