% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_db_file.R
\name{read_db_file}
\alias{read_db_file}
\title{Read annovation database file and return granges}
\usage{
read_db_file(prefix, db_path, p = 1, header = FALSE, file_encode = "UTF-8")
}
\arguments{
\item{prefix}{A prefix of database file, e.g. centromere_telomere for
hg19_centromere_telomere.txt.}

\item{db_path}{Path of database directory.}

\item{p}{A integer for the column index of chromosome, default: 1.}

\item{header}{A bool for the header of database file, default: FALSE.}

\item{file_encode}{The file endode of database file, default: UTF-8.}
}
\description{
Read annovation database file and return granges
}
