% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_cov_file.R
\name{write_cov_file}
\alias{write_cov_file}
\title{Write read count result file}
\usage{
write_cov_file(gr, path)
}
\arguments{
\item{gr}{A GRange with a column named rc.}

\item{path}{Path to write to.}
}
\description{
Write read count result file
}
